requirements.txt
fastapi
uvicorn
openai
langchain[llms]
python-dotenv
tiktoken
pytest
python-magic
selenium
beautifulsoup4
requests-mock
